added new featured section characters wildcard
