+++ characters/accessories.txt
@@ -1,0 +1,123 @@
+cross earrings
+crystal earrings
+flower earrings
+heart earrings
+hoop earrings
+multiple earrings
+pill earrings
+single earring
+skull earrings
+star earrings
+glasses
+eyewear
+monocle
+under-rim eyewear
+rimless eyewear
+semi-rimless eyewear
+red-framed eyewear
+round eyewear
+black-framed eyewear
+tinted eyewear
+medical eyepatch
+bandage over one eye
+crooked eyewear
+eyewear removed
+sunglasses
+goggles
+Blindfold
+eyepatch
+visor
+bespectacled
+blue-framed eyewear
+brown-framed eyewear
+coke-bottle glasses
+no eyewear
+opaque glasses
+over-rim eyewear
+cross necklace
+bead necklace
+pearl necklace
+heart necklace
+carrot necklace
+chain necklace
+magatama necklace
+tooth necklace
+key necklace
+anchor necklace
+skull necklace
+shell necklace
+gold necklace
+crescent necklace
+ring necklace
+feather necklace
+bone necklace
+ankh necklace
+multiple necklaces
+bullet necklace
+bracelet
+choker
+metal collar
+ring
+wristband
+pendant
+brooch
+hoop earrings
+bangle
+stud earrings
+sunburst
+pearl bracelet
+drop earrings
+puppet rings
+corsage
+sapphire brooch
+jewelry
+necklace
+ring
+wedding band
+earrings
+single earring
+stud earrings
+necklace
+jewelry
+crystal
+brooch
+gem
+chest jewel
+forehead jewel
+tassel
+belly chain 
+lace
+ribbon
+stitches
+scarf
+bandaid
+collar
+belt
+steam
+bell
+amulet
+emblem
+flag print
+anchor symbol
+cross
+diffraction spikes
+iron cross
+latin cross
+lace-trimmed hairband
+ankle lace-up
+st. gloriana's (emblem)
+heart lock (kantai collection)
+garters
+thigh strap
+thigh ribbon
+leg garter
+garter straps
+leg ribbon
+bandaid on leg
+bandaged leg
+arm garter
+joints
+kneepits
+knee pads
+thigh holster
+bandaid on knee
