+++ characters/age.txt
@@ -1,0 +1,10 @@
+Young Adult
+Adult
+Teenager
+Schoolchild
+Baby
+Toddler
+Generation Z
+20-year-old
+25 years old
+Centenarians
