+++ characters/angle.txt
@@ -1,0 +1,20 @@
+Frontal View
+Profile View
+Rear View
+Three-Quarter Rear View
+Three-Quarter View
+High Angle
+Eye Level
+Low Angle Shot
+Dutch angle
+Dramatic angle
+From above
+From behind
+From below
+From side
+Upside-Down
+Looking at viewer
+Looking at another
+Looking away
+Looking back
+Looking up
