+++ characters/artist.txt
@@ -1,0 +1,11 @@
+by Chinese Contemporary Artist
+by Contemporary Photographer
+by Contemporary Artist
+by Asian Contemporary Artist
+by European Contemporary Artists
+by American Contemporary Artists
+taken by Hideaki Hamada using Pentax 67II and Kodak Portra
+taken by Andrew Bernstein for NBA
+taken by Contemporary Artist for Harper's BAZAAR magazine
+taken by Contemporary Artist for Playboy magazine
+taken by Contemporary Artist for Vanity Fair magazine
