+++ characters/bangs.txt
@@ -1,0 +1,14 @@
+bangs
+blunt bangs
+hair between eyes
+hair over eyes
+hair over one eye
+parted bangs
+swept bangs
+sidelocks
+asymmetrical bangs
+hair intakes
+single hair intake
+Widow's Peak
+Sideburns
+Dreadlocks
