+++ characters/beard.txt
@@ -1,0 +1,20 @@
+Stubble Beard
+Goatee
+Full Beard
+Van Dyke Beard
+Soul Patch
+Garibaldi Beard
+Mutton Chops
+Circle Beard
+Corporate Beard
+Balbo Beard
+Ducktail Beard
+Chinstrap Beard
+Anchor Beard
+Chevron Mustache
+Horseshoe Mustache
+Handlebar Mustache
+Imperial Mustache
+Pencil Mustache
+Friendly Mutton Chops
+Zappa Mustache
