+++ characters/bodytype.txt
@@ -1,0 +1,19 @@
+Slim
+Fit
+Skinny
+Chubby
+Plump
+Underweight
+Normal weight
+Overweight
+Obese
+Athletic
+Muscular
+Toned
+Ripped
+Petite
+Curvy
+Fat
+abs, obliques, biceps, pectorals
+skinny, ribs, slim legs, narrow waist
+curvy, belly, stomach, thick thighs, wide hips
