+++ characters/bottomwear.txt
@@ -1,0 +1,108 @@
+Yoga shorts
+dress
+microdress
+long dress
+off-shoulder dress
+strapless dress
+backless dress
+halter dress
+sundress
+sleeveless dress
+sailor dress
+summer dress
+china dress
+pinafore dress
+sweater dress
+wedding dress
+armored dress
+frilled dress
+lace-trimmed dress
+collared dress
+fur-trimmed dress
+layered dress
+pleated dress
+taut dress
+pencil dress
+impossible dress
+multicolored dress
+striped dress
+checkered skirt
+plaid dress
+ribbed dress
+polka dot dress
+print dress
+vertical-striped dress
+see-through dress
+skirt
+microskirt
+miniskirt
+skirt suit
+bikini skirt
+pleated skirt
+pencil skirt
+bubble skirt
+tutu
+ballgown
+pettiskirt
+showgirl skirt
+Medium length skirt
+beltskirt
+denim skirt
+suspender skirt
+skirt set
+long skirt
+summer long skirt
+overskirt
+hakama skirt
+high-waist skirt
+kimono skirt
+suspender long skirt
+chiffon skirt
+frilled skirt
+fur-trimmed skirt
+lace skirt
+lace-trimmed skirt
+ribbon-trimmed skirt
+layered skirt
+print skirt
+multicolored skirt
+striped skirt
+vertical-striped skirt
+plaid skirt
+flared skirt
+floral skirt
+side slit
+shorts
+micro shorts
+short shorts
+hot pants
+cutoffs
+striped shorts
+suspender shorts
+denim shorts
+puffy shorts
+dolphin shorts
+dolfin shorts
+tight pants
+crotchless pants
+track pants
+yoga pants
+bike shorts
+gym shorts
+pants
+puffy pants
+pumpkin pants
+hakama pants
+harem pants
+bloomers
+buruma
+jeans
+cargo pants
+camouflage pants
+capri pants
+chaps
+jumpsuit
+lowleg pants
+plaid pants
+single pantsleg
+striped pants
