+++ characters/camera.txt
@@ -1,0 +1,47 @@
+Nikon D850 with Nikkor 50mm f/1.8
+Canon EOS 5D Mark IV with Canon EF 24-70mm f/2.8L II
+Sony A7R IV with Sony FE 85mm f/1.4 GM
+FujiFilm X-T4 with Fujinon XF 35mm f/2 R WR
+Olympus OM-D E-M1 Mark III with M.Zuiko 12-40mm f/2.8
+Panasonic Lumix GH5 with Leica DG 25mm f/1.4
+Leica M10 with LEICA 35mm f/2 SUMMICRON-M ASPH
+Pentax K-1 Mark II with Pentax FA 43mm f/1.9 Limited
+Sigma fp with Sigma 45mm f/2.8 DG DN
+Hasselblad X1D II with Hasselblad XCD 65mm f/2.8
+Canon EOS R with Canon RF 28-70mm f/2L
+Sony A9 II with Sony FE 24-70mm f/2.8 GM
+Nikon Z7 with Nikon Z 70-200mm f/2.8 VR S
+Fujifilm GFX 100 with GF 110mm f/2 R LM WR
+Ricoh GR III with GR 18.3mm f/2.8
+Leica Q2 with Summilux 28mm f/1.7 ASPH
+Panasonic S1R with Lumix S 50mm f/1.4
+Olympus PEN-F with M.Zuiko 17mm f/1.8
+Canon EOS M6 Mark II with Canon EF-M 32mm f/1.4
+Nikon Z50 with Nikon Z DX 16-50mm f/3.5-6.3
+Sony A6400 with Sony E 35mm f/1.8 OSS
+Fujifilm X100V with Fujinon 23mm f/2
+Pentax 645Z with Pentax-D FA 645 55mm f/2.8
+Leica SL2 with Leica APO-Summicron-SL 50mm f/2 ASPH
+Sony A1 with Sony FE 20mm f/1.8 G
+Canon EOS-1D X Mark III with Canon EF 50mm f/1.2L
+Nikon Z6 II with Nikon Z 24-70mm f/4 S
+Fujifilm X-Pro3 with Fujinon XF 56mm f/1.2 R
+Panasonic Lumix S5 with Lumix S PRO 70-200mm f/2.8 O.I.S.
+Hasselblad 907X with Hasselblad XCD 30mm f/3.5
+Olympus OM-D E-M5 Mark III with M.Zuiko 40-150mm f/2.8
+Pentax KP with Pentax HD DA 20-40mm f/2.8-4
+Sigma sd Quattro H with Sigma 24-70mm f/2.8 DG DN Art
+Ricoh Theta Z1 with Built-in 14mm f/2.1 (360-degree camera)
+Leica CL with Leica Summilux-TL 35mm f/1.4 ASPH
+Nikon D780 with Nikkor 14-24mm f/2.8G
+Canon EOS 90D with Canon EF-S 18-135mm f/3.5-5.6 IS USM
+Fujifilm X-S10 with Fujinon XF 10-24mm f/4 R OIS WR
+Sony A7C with Sony FE 28-60mm f/4-5.6
+Panasonic Lumix G9 with Leica DG 42.5mm f/1.2
+Olympus Tough TG-6 with Built-in 4.5-18mm f/2-4.9 (rugged compact)
+Kodak PIXPRO AZ901 with Built-in 4.3-258mm f/2.9-6.7 (superzoom)
+Nikon Coolpix P950 with Built-in 24-2000mm f/2.8-6.5 (superzoom)
+Canon PowerShot G5 X Mark II with Built-in 8.8-44mm f/1.8-2.8
+GoPro HERO9 with Built-in f/2.8 Ultra-Wide (action camera)
+DJI Mavic Air 2 with Built-in 24mm f/2.8 (drone camera)
+Sony RX100 VII with Built-in 24-200mm f/2.8-4.5 (compact camera)
