+++ characters/chest.txt
@@ -1,0 +1,12 @@
+Flat Chest
+Small Breasts
+Medium Breasts
+Large Breasts
+Huge Breasts
+Gigantic Breasts
+Cleavage
+Pectorals
+Collarbone
+Breasts
+Bust
+Busty
