+++ characters/color.txt
@@ -1,0 +1,24 @@
+warm color tone
+cold color tone
+pastel color tone
+vivid color tone
+vibrant color tone
+cyberpunk color tone
+steampunk color tone
+aqua
+black
+blue
+fuchsia
+gray
+green
+lime
+maroon
+navy
+olive
+orange
+purple
+red
+silver
+teal
+white
+yellow
