+++ characters/eye.txt
@@ -1,0 +1,16 @@
+Blue Eyes
+Brown Eyes
+Green Eyes
+Hazel Eyes
+Black Eyes
+Amber Eyes
+Red Eyes
+enchanting gaze
+intense gaze
+Loving Gaze
+Passionate Eyes
+Tender Look
+Admiring Glance
+Longing Look
+Gentle Eyes
+Caring Look
