+++ characters/faceexp.txt
@@ -1,0 +1,19 @@
+Happy
+Sad
+Angry
+Surprised
+Disgusted
+Fearful
+Confused
+Excited
+Tired
+Bored
+Anxious
+Content
+Determined
+Embarrassed
+Jealous
+Proud
+Shocked
+Shy
+Thoughtful
