+++ characters/faceshape.txt
@@ -1,0 +1,15 @@
+Round Face
+Oval Face
+Square Face
+Heart-Shaped Face
+Diamond Face
+Baby-Faced
+Plump Cheeks
+Chubby Cheeks
+Long Face
+Triangle Face
+Inverted Triangle Face
+Pear-shaped Face
+Oblong Face
+Square Round Face
+Square Oval Face
