+++ characters/gender.txt
@@ -1,0 +1,4 @@
+woman
+man
+female
+male
