+++ characters/gesture.txt
@@ -1,0 +1,20 @@
+Pointing
+Waving
+Thumbs Up
+Peace Sign
+Clapping
+Fist Bump
+Crossed Arms
+Hands on Hips
+Handshake
+High Five
+Salute
+Praying
+Writing
+Typing
+Holding
+Pinching
+Gesturing
+Touching
+Pushing
+Pulling
