+++ characters/haircolor.txt
@@ -1,0 +1,21 @@
+Black Hair
+Brown Hair
+Red Hair
+Blonde Hair
+Grey Hair
+White Hair
+Auburn Hair
+Chestnut Hair
+Ash Hair
+Platinum Hair
+multicolored hair
+colored inner hair
+gradient hair
+rainbow hair
+split-color hair
+streaked hair
+two-tone hair
+colored tips
+Pink Hair
+Rose Pink Gold Hair
+Chocolate Brown Hair
