+++ characters/hairstyle.txt
@@ -1,0 +1,38 @@
+Waist-length hair
+Shoulder-length hair
+Long hair
+Medium-length hair
+Short hair
+Long Straight Hair
+Ponytail
+Twintails
+Bun
+Bob Cut
+Pixie Cut
+Layered Cut
+Updo
+Braid
+Half Updo
+Beach Waves
+Long Curls
+Crew Cut
+Buzz Cut
+Side Part
+Textured Quiff
+Spiky Hair
+Slicked Back
+Taper Fade
+Pompadour
+Undercut
+Bowl Cut
+Faux Hawk
+Cornrows
+Mohawk
+Afro
+high ponytail
+low ponytail
+side ponytail
+short ponytail
+folded ponytail
+braided ponytail
+topknot
