+++ characters/job.txt
@@ -1,0 +1,57 @@
+race queen
+queen
+maid
+housemaid
+wa maid
+office lady
+waitress
+actress
+dancing girl
+stewardess
+nun
+dancer
+ballerina
+gym leader
+idol
+lifeguard
+boxer
+scientist
+athletes
+monk
+crobat
+nurse
+student
+teacher
+racer
+police
+soldier
+cheerleader
+actor
+motorcyclist
+hacker
+magician
+detective
+pilot
+diver
+bar censor
+missionary
+firefighter
+goalkeeper
+chef
+astronaut
+cashier
+mailman
+barista
+doctor
+trap
+spy
+agent
+assassin
+poet
+samurai
+the hermit
+makihitsuji
+knight
+priest
+cleric
+doll
