+++ characters/legs.txt
@@ -1,0 +1,7 @@
+Thick Thighs
+Long Legs
+Slim Legs
+Bare foot
+Thighs
+Legs
+Feet
