+++ characters/lighting.txt
@@ -1,0 +1,35 @@
+Cinematic lighting
+Dramatic light
+Golden hour light
+Studio lighting
+Available light
+Natural sunlight
+Soft ambient light
+Harsh sunlight
+Overcast sky
+Sunset glow
+Sunrise warmth
+Twilight hues
+Candlelight
+Incandescent lighting
+Fluorescent lighting
+Moonlight
+Dappled sunlight
+Backlit silhouette
+Spotlight
+Rim lighting
+Firelight
+City streetlights
+Lantern light
+Tungsten lighting
+Cloudy day diffused light
+Skylight
+Blue hour light
+Flash photography
+Stage lighting
+Neon lights
+Torchlight
+Softbox lighting
+Rim light
+Lightning
+Abstract light patterns
