+++ characters/otherfeatures.txt
@@ -1,0 +1,13 @@
+Dimple
+Freckles
+Wrinkles
+Mole
+Scar
+High Cheek Bones
+makeup
+lipstick
+thick eyebrows
+mole under eye
+mole under mouth
+scar on face
+facepaint
