+++ characters/pose.txt
@@ -1,0 +1,37 @@
+Dramatic Pose
+Standing Pose
+Sitting Pose
+Leaning Pose
+Lying Pose
+Running Pose
+Jumping Pose
+Dancing Pose
+Yoga Pose
+Fighting Pose
+Crossed Arms Pose
+Hands on Hips Pose
+Pointing Pose
+Waving Pose
+Praying Pose
+Reading Pose
+Writing Pose
+Cooking Pose
+Driving Pose
+Swimming Pose
+Flying Pose
+Power Pose
+Walking Pose
+The Over-the-Shoulder Look
+S-curve Pose
+Close-Up Beauty Shot Pose
+Arms Up Pose
+Casual Stroll Pose
+Headshot Pose
+Sitting Cross-Legged Pose
+Back Arch Pose
+Hand-on-Hip Pose
+Gazing into the Distance Pose
+Candid Laugh Pose
+Dynamic Action Pose
+Contrapposto Pose
+High Fashion Pose
