+++ characters/quality.txt
@@ -1,0 +1,14 @@
+HDR
+UHD
+highly detailed
+natural skin
+sharp focus
+sharp details
+intricate details
+aesthetic
+photogenic
+gorgeous
+perfect
+HDR, UHD, (8k resolution:1.3), 64K, (highly detailed:1.2), natural skin, sharp focus, sharp details, beautiful details, aesthetic, photogenic, gorgeous, perfect
+depth of field
+bokeh
