+++ characters/region.txt
@@ -1,0 +1,22 @@
+From China
+From Eastern Europe
+From East Asia
+From Japan
+From South Korea
+From Thailand
+From Vietnam
+From Russia
+From Belarus
+From the United States
+From England
+From Africa
+From Europe
+From South America
+From Taiwan
+From Heilongjiang Heihe
+From Urumqi, Xinjiang
+From Sanya, Hainan
+From Chengdu, Sichuan
+From Yunnan
+From Shanghai
+From Jiangnan
