+++ characters/shoes.txt
@@ -1,0 +1,60 @@
+barefoot
+no shoes
+shoes removed
+single shoe
+the only shoe
+black loafers
+shoes
+uwabaki
+platform footwear
+high heels
+stiletto heels
+strappy heels
+platform heels 
+rudder footwear
+sandals
+barefoot sandals
+clog sandals
+geta
+slippers
+skates
+roller skates
+animal feet
+animal slippers
+anklet
+shackles
+brown footwear
+cross-laced footwear
+loafers
+mary janes
+mismatched footwear
+pointy footwear
+pumps
+sandals removed
+shoe soles
+shoelaces
+sneakers
+winged footwear
+zouri
+boots
+boots removed
+thigh boots
+knee boots
+cross-laced footwear
+ankle boots
+high heel boots
+toeless boots
+lace-up boots
+fur-trimmed boots
+snow boots
+anklet
+rubber boots
+santa boots
+leather boots
+thighhighs under boots
+combat boots
+doc martens
+rain boots
+belt boots
+fur boots
+single boot
