+++ characters/shot.txt
@@ -1,0 +1,15 @@
+Portrait
+Close Up Shot
+Full Shot
+Waist Shot
+Medium Shot
+Macro Shot
+Full Body View
+Upper Body View
+Product View
+Aerial View
+Outer Space View
+Extreme Close Up Shot
+Tilt-shift Photography
+Fish-eye Lens Photo
+Over-the-Shoulder Shot
