+++ characters/socks.txt
@@ -1,0 +1,50 @@
+bodystocking
+pantyhose
+leggings
+legwear
+thighhighs
+kneehighs
+socks
+bare legs
+black bodystocking
+white bodystocking
+stocking under clothes
+black pantyhose
+white pantyhose
+thighband pantyhose
+pantylines
+single leg pantyhose
+panties under pantyhose
+fishnets
+stirrup legwear
+toeless legwear
+mismatched legwear
+two-tone legwear
+asymmetrical legwear
+uneven legwear
+white thighhighs
+black thighhighs
+pink thighhighs
+suspenders
+garter straps
+torn legwear
+torn thighhighs
+see-through legwear
+frilled legwear
+lace-trimmed legwear
+seamed legwear
+back-seamed legwear
+animal ear legwear
+striped legwear
+vertical-striped legwear
+polka dot legwear
+print legwear
+legwear under shorts
+over-kneehighs
+bobby socks
+tabi
+loose socks
+ankle socks
+leg warmers
+single sock
+striped socks
