+++ characters/suit.txt
@@ -1,0 +1,41 @@
+suit
+tuxedo
+formal dress
+evening gown
+canonicals
+cocktail dress
+gown
+sash
+cheongsam
+china dress
+print cheongsam
+pelvic curtain
+wedding dress
+uchikake
+school uniform
+sailor
+serafuku
+summer uniform
+kindergarten uniform
+police uniform
+naval uniform
+military uniform
+ss uniform/nazi uniform
+maid
+stile uniform
+miko
+overalls
+business suit
+nurse
+chef uniform
+labcoat
+cheerleader
+band uniform
+space suit
+leotard
+domineering
+japanese clothes
+kimono
+sleeveless kimono
+short kimono
+print kimono
