+++ characters/topwear.txt
@@ -1,0 +1,43 @@
+Yoga vest
+blouse
+white shirt
+collared shirt
+dress shirt
+sailor shirt
+cropped shirt
+t-shirt
+casual T-shirts
+short sleeve T-shirts
+off-shoulder shirt
+shrug (clothing)
+cardigan
+criss-cross halter
+frilled shirt
+sweatshirt
+hawaiian shirt
+hoodie
+Impossible shirt
+kappougi
+plaid shirt
+polo shirt
+print shirt
+shirt
+sleeveless hoodie
+sleeveless shirt
+striped shirt
+tank top
+vest
+waistcoat
+camisole
+tied shirt
+undershirt
+crop top
+highleg
+clothing cutout
+back cutout
+cleavage cutout
+navel cutout
+midriff
+heart cutout
+torn clothes
+torn shirt
