+++ characters/waist.txt
@@ -1,0 +1,10 @@
+Narrow Waist
+Wide Hips
+Abs
+Belly
+Midriff
+Stomach
+Groin
+Navel
+Waist
+Pregnant
