+++ modules/config.py
@@ -233,6 +233,11 @@ default_advanced_checkbox = get_config_item_or_set_default(
     default_value=False,
     validator=lambda x: isinstance(x, bool)
 )
+default_character_checkbox = get_config_item_or_set_default(
+    key='default_character_checkbox',
+    default_value=False,
+    validator=lambda x: isinstance(x, bool)
+)
 default_max_image_number = get_config_item_or_set_default(
     key='default_max_image_number',
     default_value=1000,
